Refactored crypto providers to be split up in cipher, signature and mac providers
