--- rxkeystore/src/androidTest/java/com/nexenio/rxkeystore/provider/asymmetric/ec/RxECCryptoProviderTest.java
@@ -1,35 +1,0 @@
-package com.nexenio.rxkeystore.provider.asymmetric.ec;
-
-import com.nexenio.rxkeystore.RxKeyStore;
-import com.nexenio.rxkeystore.provider.asymmetric.BaseAsymmetricCryptoProviderTest;
-import com.nexenio.rxkeystore.provider.asymmetric.RxAsymmetricCryptoProvider;
-
-import org.junit.Before;
-import org.junit.Ignore;
-import org.junit.Test;
-
-import androidx.annotation.CallSuper;
-import androidx.annotation.NonNull;
-
-public class RxECCryptoProviderTest extends BaseAsymmetricCryptoProviderTest {
-
-    @CallSuper
-    @Before
-    @Override
-    public void setUpBeforeEachTest() {
-        super.setUpBeforeEachTest();
-    }
-
-    @Override
-    protected RxAsymmetricCryptoProvider createAsymmetricCryptoProvider(@NonNull RxKeyStore keyStore) {
-        return new RxECCryptoProvider(keyStore);
-    }
-
-    @Ignore("Unclear if this is the expected behaviour")
-    @Test
-    @Override
-    public void sign_subsequentCallsWithSameKey_emitsSameSignature() {
-        super.sign_subsequentCallsWithSameKey_emitsSameSignature();
-    }
-
-}
